Fix bug where bot was skipping RuneScape messages
When the bot restarts, every few hours the messages are cleared in the
RuneScape chat. This change would not be recognized by the bot so it
would end up waiting for message 47 for example, but there are actually
only 3 messages available at the time, so it would be waiting a while
until it actually prints something to Discord.

This has been fixed, satisfying Issue #8 in GitHub.