Start to comply with the AGPL-3.0
Right now it only prints a message to Discord that links the GitHub
page. I want a message to also be sent in RuneScape and I want the bot
to handle something like a "license" command that explains the license
and links the source code for users in both RuneScape chat and Discord
chat.