Make the bot OOP so that people can build upon it
The bot is now split into two distinct classes (DiscordSync and
RuneScapeSync) that each control the chat for their own service. main.js
manages the two, and Queue is a new datatype used to hold the message
data and the author data (and soon maybe timestamps as well). The design
still needs to be cleaned up a little bit, but I wanted to get a release
out.

This should satisfy Issue #13 on GitHub.