Add pause and resume commands
pause <service> and resume <service> have just been added as commands to
the console interface, with "runescape", "discord", or "both" being
acceptable values for "service".

The bot is nearing completion. Now it just needs an updated config
system to match the new feature additions, and the code needs to be
cleaned up so that it's more readable.