Fix screenshot and html commands
Apparently I still don't know how promises work so the screenshot and
html commands just returned "[object Promise]". This has been fixed.
Hopefully in the future I'll find a better way to fix this since I don't
like making every function async, but for now this is how it is.

Note that this commit is signed with a new PGP key. I intend to use this
key as my primary key from now on.