Fix random Discord crashes
Sometimes discord.js will unexpectedly throw errors, requiring a restart
of the Discord part of the bot. This has been fixed, resolving Issue #15
on GitHub.

Note that this commit is being signed with my old PGP key that is being
replaced by the one that signed the last commit. This will be the last
commit in this repository that is signed with this key.